+++ samples/Java/TestJavaClient/SampleFrame.java
@@ -1195,7 +1195,7 @@ class SampleFrame extends JFrame implements EWrapper {
         m_tickers.add(msg);
     }
 
-    public void orderStatus( int orderId, String status, double filled, double remaining,
+    public void orderStatus( int orderId, String status, Decimal filled, Decimal remaining,
     						 double avgFillPrice, int permId, int parentId,
     						 double lastFillPrice, int clientId, String whyHeld, double mktCapPrice) {
         // received order status
@@ -1369,7 +1369,7 @@ class SampleFrame extends JFrame implements EWrapper {
         m_acctDlg.updateAccountValue(key, value, currency, accountName);
     }
 
-    public void updatePortfolio(Contract contract, double position, double marketPrice,
+    public void updatePortfolio(Contract contract, Decimal position, double marketPrice,
         double marketValue, double averageCost, double unrealizedPNL, double realizedPNL,
         String accountName) {
         m_acctDlg.updatePortfolio(contract, position, marketPrice, marketValue,
@@ -1563,7 +1563,7 @@ class SampleFrame extends JFrame implements EWrapper {
         destOrder.autoCancelParent(srcOrder.autoCancelParent());
     }
 
-    public void position(String account, Contract contract, double pos, double avgCost) {
+    public void position(String account, Contract contract, Decimal pos, double avgCost) {
         String msg = EWrapperMsgGenerator.position(account, contract, pos, avgCost);
         m_TWS.add(msg);
     }
@@ -1583,7 +1583,7 @@ class SampleFrame extends JFrame implements EWrapper {
         m_TWS.add(msg);
     }
 
-    public void positionMulti( int reqId, String account, String modelCode, Contract contract, double pos, double avgCost) {
+    public void positionMulti( int reqId, String account, String modelCode, Contract contract, Decimal pos, double avgCost) {
         String msg = EWrapperMsgGenerator.positionMulti(reqId, account, modelCode, contract, pos, avgCost);
         m_TWS.add(msg);
     }
@@ -1761,7 +1761,7 @@ class SampleFrame extends JFrame implements EWrapper {
     }
 
     @Override
-    public void pnlSingle(int reqId, int pos, double dailyPnL, double unrealizedPnL, double realizedPnL, double value) {
+    public void pnlSingle(int reqId, Decimal pos, double dailyPnL, double unrealizedPnL, double realizedPnL, double value) {
         String msg = EWrapperMsgGenerator.pnlSingle(reqId, pos, dailyPnL, unrealizedPnL, realizedPnL, value);
         
         m_TWS.add(msg);
